Add auth-data-watch-oauth module
