+++ auth-data-watch-oauth/README
@@ -1,0 +1,17 @@
+# Auth Data Watch OAuth library
+
+[![Maven Central](https://img.shields.io/maven-central/v/com.google.android.horologist/horologist-auth-data-watch-oauth)](https://search.maven.org/search?q=g:com.google.android.horologist)
+
+For more information, visit the documentation: https://google.github.io/horologist/auth-data-watch-oauth
+
+## Download
+
+```groovy
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    implementation "com.google.android.horologist:horologist-auth-data-watch-oauth:<version>"
+}
+```
