+++ auth-data-watch-oauth/gradle.properties
@@ -1,0 +1,3 @@
+POM_ARTIFACT_ID=horologist-auth-data-watch-oauth
+POM_NAME=Horologist Auth Data Watch OAuth library
+POM_PACKAGING=aar
