--- auth-data/src/main/java/com/google/android/horologist/auth/data/oauth/common/impl/google/api/TokenResponse.kt
@@ -1,30 +1,0 @@
-/*
- * Copyright 2022 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.google.android.horologist.auth.data.oauth.common.impl.google.api
-
-import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-public data class TokenResponse(
-    @Json(name = "access_token") val accessToken: String,
-    @Json(name = "expires_in") val expiresIn: Int,
-    @Json(name = "id_token") val idToken: String?,
-    @Json(name = "refresh_token") val refreshToken: String?,
-    @Json(name = "scope") val scope: String?,
-    @Json(name = "token_type") val tokenType: String
-)
