+++ settings.gradle.kts
@@ -37,6 +37,7 @@ include(":audio-ui")
 include(":auth-composables")
 include(":auth-data")
 include(":auth-data-phone")
+include(":auth-data-watch-oauth")
 include(":auth-sample-phone")
 include(":auth-sample-shared")
 include(":auth-sample-wear")
